Implement transaction custom view, adapter, partial implementation of database interactions, add currentTransactionActivity

+++ app/src/main/java/com/example/projet_pfe_android/Util/JavaUtil.java
@@ -5,7 +5,7 @@ import java.text.NumberFormat;
 public class JavaUtil {
 
 //    Gives proper format to currency values for display
-    public static String currencyString(float value){
+    public static String currencyString(double value){
         return NumberFormat.getCurrencyInstance().format(value);
     }
 }
